src/training/requirements.txt
loguru==0.7.0
pandas>=1.3.0
psycopg2-binary==2.9.10
ray[data,serve,train,tune]>=2.39.0
xgboost-ray==0.1.19
